Backend/pyproject.toml
[project]
name = "BuildXvibe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "e2b-code-interpreter>=2.3.0",
    "fastapi[standard]>=0.121.1",
    "langgraph>=1.0.2",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
]
